requirements.txt
langchain
python-dotenv
langchain-community
langchain-groq
PyPDF2 
PyMuPDF 